% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ff.R
\name{as.ram2}
\alias{as.ram2}
\title{as.ram2
wrapper for ff:as.ram that also deletes the underlying ff/ffdf to reduce temp space}
\usage{
as.ram2(obj)
}
\arguments{
\item{obj}{ff/ffdf object to be loaded to ram and deleted}
}
\value{
a data.table
}
\description{
as.ram2
wrapper for ff:as.ram that also deletes the underlying ff/ffdf to reduce temp space
}
\examples{

library(ff)

test <- ff(1:5)
file.exists(filename(test)) # TRUE
test.ram <- as.ram2(test)
file.exists(filename(test)) # FALSE

}
