% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_create}
\alias{cache_create}
\title{cache_create}
\usage{
cache_create(
  table_name,
  depth = c("deep", "shallow"),
  type = c("tessi", "stream")
)
}
\arguments{
\item{table_name}{string}

\item{depth}{string, either "deep" or "shallow"}

\item{type}{string, either "tessi" or "stream"}
}
\value{
string for the configured cache path
}
\description{
Internal function to create cache directory, simply defined by
\verb{tessilake.\{depth\}/\{type\}}
}
\examples{
\dontrun{
cache_create("test","deep","stream")
}
}
