% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_files.R
\name{cache_delete}
\alias{cache_delete}
\title{cache_delete}
\usage{
cache_delete(table_name, depth, type, partitions = NULL)
}
\arguments{
\item{table_name}{string}

\item{depth}{string, e.g. "deep" or "shallow"}

\item{type}{string, e.g. "tessi" or "stream"}

\item{partitions}{optional vector of partitions to delete}
}
\value{
nothing, invisibly
}
\description{
Delete some or all cache files for a given cache.
}
\examples{
\dontrun{
cache_delete("test", "deep", "stream", partitions = c(1, 2, 3))
}
}
