% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_files.R
\name{cache_files}
\alias{cache_files}
\title{cache_files}
\usage{
cache_files(table_name, depth, type)
}
\arguments{
\item{table_name}{string}

\item{depth}{string, e.g. "deep" or "shallow"}

\item{type}{string, e.g. "tessi" or "stream"}
}
\value{
character vector of filenames
}
\description{
Get a vector of all of the files that make up the cache
}
