% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_files.R
\name{cache_get_mtime}
\alias{cache_get_mtime}
\title{cache_get_mtime}
\usage{
cache_get_mtime(table_name, depth, type)
}
\arguments{
\item{table_name}{string}

\item{depth}{string, e.g. "deep" or "shallow"}

\item{type}{string, e.g. "tessi" or "stream"}
}
\value{
POSIXct modification time
}
\description{
Get the last modification time of the cache by querying the \code{mtime} of all of the files that make up the cache
}
\examples{
\dontrun{
cache_get_mtime("test", "deep", "stream")
}
}
