% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{cache_get_partitioning}
\alias{cache_get_partitioning}
\title{cache_get_partitioning}
\usage{
cache_get_partitioning(x)
}
\arguments{
\item{x}{Arrow Dataset}
}
\value{
string column name of partition
}
\description{
Internal function to read/generate partition information from an Arrow Dataset.
}
