% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{cache_make_partitioning}
\alias{cache_make_partitioning}
\title{cache_make_partitioning}
\usage{
cache_make_partitioning(x, primaryKeys = attr(x, "primaryKeys"))
}
\arguments{
\item{x}{data.frame}

\item{primaryKeys}{primary keys to set partitioning with}
}
\value{
un-evaluated R expression to be used in creating the partitioning by
}
\description{
Builds suggested partitioning from the primary keys of the table.
}
\examples{
x = data.frame(a=c(1,2,3),b=c("a","b","c"))

cache_make_partitioning(x,primaryKeys="a")
# expression a/10000

x \%>\% mutate(partition = eval(cache_make_partitioning(x,primaryKeys="a")))

}
