% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{cache_read}
\alias{cache_read}
\title{cache_read}
\usage{
cache_read(
  tableName,
  depth = c("deep", "shallow"),
  type = c("tessi", "stream"),
  include.partition = F,
  ...
)
}
\arguments{
\item{tableName}{string}

\item{depth}{string, either "deep" or "shallow"}

\item{type}{string, either "tessi" or "stream"}

\item{include.partition}{boolean, whether or not to return the partition information as a column}
}
\value{
\code{\link[arrow:Dataset]{arrow::Dataset}}
}
\description{
Internal function to read cached arrow files. Simple wrapper on open_dataset that points to a directory defined by
\verb{tessilake.\{depth\}/\{type\}}. Optionally returns the partition information.
}
