% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{cache_update}
\alias{cache_update}
\title{cache_update}
\usage{
cache_update(
  x,
  tableName,
  depth = c("deep", "shallow"),
  type = c("tessi", "stream"),
  primaryKeys = attr(x, "primaryKeys"),
  ...
)
}
\arguments{
\item{x}{data.frame or part of a data.frame to be cached}

\item{tableName}{string}

\item{depth}{string, either "deep" or "shallow"}

\item{type}{string, either "tessi" or "stream"}
}
\value{
invisible
}
\description{
Internal function to update portions of cached arrow files.
}
