% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{cache_write}
\alias{cache_write}
\title{cache_write}
\usage{
cache_write(
  x,
  tableName,
  depth = c("deep", "shallow"),
  type = c("tessi", "stream"),
  primaryKeys = attr(x, "primaryKeys"),
  ...
)
}
\arguments{
\item{x}{data.frame to be written}

\item{tableName}{string}

\item{depth}{string, either "deep" or "shallow"}

\item{type}{string, either "tessi" or "stream"}

\item{primaryKeys}{character vector of columns to be used for partitioning, only the first one is currently used}
}
\value{
invisible
}
\description{
Internal function to write cached arrow files. Simple wrapper on write_dataset that points to a directory defined by
\verb{tessilake.\{depth\}/\{type\}} and uses partitioning specified by primaryKeys attribute/argument.
}
