% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_table.R
\name{expr_get_names}
\alias{expr_get_names}
\title{expr_get_names}
\usage{
expr_get_names(expr)
}
\arguments{
\item{expr}{expression to extract names from}
}
\value{
vector of strings, the names referred to in the expression
}
\description{
Extract (symbol) names from an expression. The expression must be a list or vector of strings and/or symbols
}
\examples{
expr_get_names(expr("a")) == expr_get_names(expr(a))
expr_get_names(expr(c("a", b)))
# c("a","b")
expr_get_names(expr(list("a", b = c)))
# c("a",b="c")

}
