% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ff.R
\name{fix_vmode}
\alias{fix_vmode}
\title{fix_vmode
adjust the vmode of a vector to the lowest necessary to save it as ff}
\usage{
fix_vmode(vec)
}
\arguments{
\item{vec}{the vector to be adapted}
}
\value{
a vector with the vmode attribute set appropriately
}
\description{
fix_vmode
adjust the vmode of a vector to the lowest necessary to save it as ff
}
\examples{

library(ff)
test <- 1:5
test <- fix_vmode(test)
vmode(test) # "nibble"
test <- fix_vmode(test[1:3])
vmode(test) # "quad"

}
