% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessi.R
\name{is.error}
\alias{is.error}
\title{is.error
helper function to use errorable expressions in stopifnot}
\usage{
is.error(expr)
}
\arguments{
\item{expr}{expression to run (in the parent frame)}
}
\value{
boolean, TRUE if the expression causes an error, FALSE if not
}
\description{
is.error
helper function to use errorable expressions in stopifnot
}
