% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_error.R
\name{is_error}
\alias{is_error}
\title{is_error}
\usage{
is_error(expr)
}
\arguments{
\item{expr}{expression to run (in the parent frame)}
}
\value{
boolean, TRUE if the expression causes an error, FALSE if not
}
\description{
Helper function to use error-throwing expressions in tests
}
\examples{
is_error(stop("Error"))

}
