% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{local_cache_dirs}
\alias{local_cache_dirs}
\title{local_cache_dirs}
\usage{
local_cache_dirs(..., envir = parent.frame())
}
\arguments{
\item{...}{other arguments for compatibility with testthat?}

\item{envir}{parent.frame()}
}
\description{
Test helper, create cache directories for tests.
}
