% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessi.R
\name{merge_customer_no_map}
\alias{merge_customer_no_map}
\title{merge_customer_no_map}
\usage{
merge_customer_no_map(
  table,
  column,
  freshness = as.difftime(7, units = "days")
)
}
\arguments{
\item{table}{data.table to merge with tessi_customer_no_map}

\item{column}{string column name to merge with}

\item{freshness}{difftime}
}
\value{
merged data.table
}
\description{
merge_customer_no_map
}
