% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessi.R
\name{read_cache}
\alias{read_cache}
\title{read_tessi_cache
internal function to read cached tessi files}
\usage{
read_cache(
  tableName,
  type = c("tessi", "stream"),
  depth = c("deep", "shallow")
)
}
\arguments{
\item{tableName}{string}

\item{type}{string, either "tessi" or "stream"}

\item{depth}{string, either "deep" or "shallow"}
}
\value{
environment containing the deep and shallow caches as arrow Datasets
}
\description{
read_tessi_cache
internal function to read cached tessi files
}
