% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{read_sql}
\alias{read_sql}
\title{read_sql}
\usage{
read_sql(
  query,
  name = digest::sha1(query),
  primary_keys = NULL,
  date_column = NULL,
  freshness = as.difftime(7, units = "days")
)
}
\arguments{
\item{query}{character query to run on the database.}

\item{name}{name of the query, defaults to the SHA1 hash of the query.}

\item{primary_keys}{primary keys of the query to be used for incremental updates}

\item{date_column}{update date column of the query to be used for incremental updates}

\item{freshness}{the returned data will be at least this fresh}
}
\value{
an Apache Arrow Table, see the \link[arrow:arrow-package]{arrow::arrow-package} package for more information.
}
\description{
Execute a database query and cache it locally as a Feather file and remotely as a Parquet file.
Cache storage locations are managed by \code{tessilake.shallow} and \code{tessilake.deep} options.
Database connection defined by an ODBC profile with the name set by the \code{tessilake.tessitura} option.
}
