% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tessiriver.R
\name{read_stream}
\alias{read_stream}
\title{read_stream}
\usage{
read_stream(streamName, subset, select, ...)
}
\arguments{
\item{streamName}{character name of the stream file, with or without the extension}

\item{subset}{logical expression indicating elements or rows to keep}

\item{select}{character vector indicating columns to select from stream file}

\item{...}{further arguments to be passed to or from other methods}
}
\value{
a data.table of the subsetted stream, invisibly
}
\description{
read_stream
}
\examples{

if(FALSE){read_stream("addressStream",
         timestamp>=as.Date("2021-07-01"),
         c("group_customer_no","timestamp")}


}
