% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessi.R
\name{read_tessi}
\alias{read_tessi}
\title{read_tessi}
\usage{
read_tessi(tableName, subset, select, ...)
}
\arguments{
\item{tableName}{character name of the table to read from Tessitura, either one of the available tables (see "list_tessi_tables") or a
table that exists in Tessitura. The default schema is "dbo"}

\item{subset}{logical expression indicating elements or rows to keep}

\item{select}{character vector indicating columns to select from stream file}

\item{...}{further arguments to be passed to or from other methods}
}
\value{
an Apache Arrow Table, see the "arrow" package for more information
}
\description{
Read a table from Tessitura and cache it locally as a Parquet file.
Cache storage is managed by "tessilake.shallow" option.
Tessitura database connection defined by an ODBC profile with the name set by the "tessilake.Tessitura" option.
}
\examples{

\dontrun{read_tessi("memberships",
         init_dt>=as.Date("2021-07-01"),
         c("memb_level","customer_no")}

}
