% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessi.R
\name{read_tessi_customer_no_map}
\alias{read_tessi_customer_no_map}
\title{read_tessi_customer_no_map}
\usage{
read_tessi_customer_no_map(freshness = as.difftime(7, units = "days"))
}
\arguments{
\item{freshness}{the returned data will be at least this fresh}
}
\value{
\link[arrow:Table]{arrow::Table}
}
\description{
Return an Arrow table of customer numbers \code{customer_no_old} mapped to merged \code{customer_no} and household/primary group
customer numbers \code{group_customer_no}.
}
