% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessi.R
\name{read_tessi_db}
\alias{read_tessi_db}
\title{read_tessi_db
internal function to return a Tessitura table based on a name}
\usage{
read_tessi_db(tableName)
}
\arguments{
\item{tableName}{string}
}
\value{
dplyr database query
}
\description{
read_tessi_db
internal function to return a Tessitura table based on a name
}
\examples{

}
