% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setleftjoin.R
\name{setleftjoin}
\alias{setleftjoin}
\title{setleftjoin
Fast left join for data.table that simply adds columns from \code{r} to \code{l} by reference.
Requires that \code{r} has at most one item of each join group in order to be able to simply add to \code{l}.}
\usage{
setleftjoin(l, r, by = NULL)
}
\arguments{
\item{l}{left side of the join (data.table)}

\item{r}{right side of the join (data.table)}

\item{by}{character vector for matching columns. Names, if they exist, refer to columns in the
left table, values to columns in the right table}
}
\value{
a joined data.table, invisibly
}
\description{
setleftjoin
Fast left join for data.table that simply adds columns from \code{r} to \code{l} by reference.
Requires that \code{r} has at most one item of each join group in order to be able to simply add to \code{l}.
}
\examples{

library(data.table)

l <- data.table(a = c(1, 2, 3), b = c(3, 4, NA))
r <- data.table(b = c(3, 7, NA), c = c(1, 2, 4))

setleftjoin(l, r, by = "b")
setleftjoin(l, r, by = c("b" = "c"))

}
