% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql_connect}
\alias{sql_connect}
\alias{sql_disconnect}
\title{sql_connect}
\usage{
sql_connect()

sql_disconnect()
}
\value{
invisible NULL
}
\description{
Connects and disconnects from the SQL database using the \code{tessilake.tessitura} connection string in config.yml and
updates the Rstudio connection pane
}
\section{Functions}{
\itemize{
\item \code{sql_disconnect()}: Tear down the SQL connection

}}
\examples{
\dontrun{
tessilake:::db$db
sql_connect()
tessilake:::db$db
sql_disconnect()
}
}
