% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessi.R
\docType{data}
\name{tessi_list_tables}
\alias{tessi_list_tables}
\alias{tessiTables}
\title{tessi_list_tables}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 52 rows and 4 columns.
}
\usage{
tessi_list_tables()

tessiTables
}
\value{
data.table of configured tessitura tables with columns shortName, longName, baseTable and primaryKeys
}
\description{
The list of tessitura tables is configured in the extdata/tessiTables.yml file in the package directory
}
\details{
\subsection{yml format}{\preformatted{\{shortName\}:
   longName: \{name of table/view to be loaded\}
   baseTable: \{the underlying table being queried that has primaryKeys\}
   primaryKeys: \{the primary key(s) as a value or a list of values\}
}
}
}
\examples{
#customers:
#   longName: BI.VT_CUSTOMER
#   baseTable: T_CUSTOMER
#   primaryKeys: customer_no

list_tessi_tables()[shortName=="customers"]

}
\keyword{datasets}
