% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessi.R
\name{tessi_read_db}
\alias{tessi_read_db}
\title{tessi_read_db
internal function to return a Tessitura table based on a name}
\usage{
tessi_read_db(table_name)
}
\arguments{
\item{table_name}{string}
}
\value{
dplyr database query
}
\description{
tessi_read_db
internal function to return a Tessitura table based on a name
}
\examples{
\dontrun{
tessi_read_db("seasons")
}
}
