% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessi.R
\name{update_data.table}
\alias{update_data.table}
\title{update_data.table
updates the data.table-like object "to" based on the data.frame-like "from" object,
using "dateColumn" to determine which rows have been updated
and updating only those with matching "primaryKey".}
\usage{
update_data.table(from, to, dateColumn, primaryKeys)
}
\arguments{
\item{from}{the source object}

\item{to}{the object to be updated}

\item{dateColumn}{string or tidyselected column identifying the date column to use to determine which rows to update}

\item{primaryKeys}{vector of strings or tidyselected columns identifying the primary keys to determine which rows to update}
}
\value{
an updated data.table updated in-place
}
\description{
update_data.table
updates the data.table-like object "to" based on the data.frame-like "from" object,
using "dateColumn" to determine which rows have been updated
and updating only those with matching "primaryKey".
}
