% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_table.R
\name{update_table}
\alias{update_table}
\title{update_table}
\usage{
update_table(from, to, date_column = NULL, primary_keys = NULL)
}
\arguments{
\item{from}{the source object}

\item{to}{the object to be updated}

\item{date_column}{string or tidyselected column identifying the date column to use to determine which rows to update}

\item{primary_keys}{vector of strings or tidyselected columns identifying the primary keys to determine which rows to update}
}
\value{
an updated data.table updated in-place
}
\description{
updates the data.table-like object \code{to} based on the data.frame-like \code{from} object,
using \code{date_column} to determine which rows have been updated and updating only those with matching \code{primary_keys}.
}
\examples{
library(data.table)
expect <- data.table(expand.grid(x = 1:100, y = 1:100))[, data := runif(.N)]
# divide the data
from <- copy(expect)[1:9000]
to <- copy(expect)[1000:10000]
# and mung it up
to[1:5000, data := runif(.N)]

update_table(from, to, primary_keys = c(x,y)) == expect
# TRUE
}
