% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_readwrite.R
\name{write_cache}
\alias{write_cache}
\title{write_cache}
\usage{
write_cache(
  x,
  table_name,
  depth,
  type,
  primary_keys = cache_get_attributes(x)$primary_keys,
  partition = !is.null(primary_keys),
  overwrite = FALSE,
  num_tries = 60,
  ...
)
}
\arguments{
\item{x}{data.frame to be written}

\item{table_name}{string}

\item{depth}{string, e.g. "deep" or "shallow"}

\item{type}{string, e.g. "tessi" or "stream"}

\item{primary_keys}{character vector of columns to be used for partitioning, only the first one is currently used}

\item{partition}{boolean, whether or not to partition the dataset using primary_keys information}

\item{overwrite}{boolean, whether or not to overwrite an existing cache}

\item{num_tries}{integer number of times to try reading before failing}

\item{...}{extra arguments passed on to \link[arrow:write_feather]{arrow::write_feather}, \link[arrow:write_parquet]{arrow::write_parquet} or \link[arrow:write_dataset]{arrow::write_dataset}}
}
\value{
invisible
}
\description{
Function to write cached arrow files. Simple wrapper on write_dataset that points to a directory defined by
\verb{tessilake.\{depth\}/\{type\}} and uses partitioning specified by primary_keys attribute/argument.
}
\examples{
\dontrun{
x <- data.table(a = c(1, 2, 3))
write_cache(x, "test", "deep", "stream", primary_keys = c("a"))
}
}
