% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_pathexists.R
\name{cache_exists}
\alias{cache_exists}
\title{cache_exists}
\usage{
cache_exists(
  table_name,
  depth = c("deep", "shallow"),
  type = c("tessi", "stream")
)
}
\arguments{
\item{table_name}{string}

\item{depth}{string, either "deep" or "shallow"}

\item{type}{string, either "tessi" or "stream"}
}
\value{
TRUE/FALSE
}
\description{
Internal function to test if a cache already exists.
}
\examples{
\dontrun{
cache_exists("test", "deep", "stream")
}
}
