% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_attributes.R
\name{cache_get_attributes}
\alias{cache_get_attributes}
\title{cache_get_attributes}
\usage{
cache_get_attributes(x)
}
\arguments{
\item{x}{\link[arrow:arrow-package]{arrow} Object}
}
\value{
a named vector of attributes
}
\description{
Internal function to read R attributes on an \link[arrow:Dataset]{arrow::Dataset}/\code{arrow:::arrow_dplyr_query} or other object
}
\examples{
\dontrun{
cache_get_attributes(cache_read("test", "deep", "stream"))
}
}
