% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_get_partitioning}
\alias{cache_get_partitioning}
\title{cache_get_partitioning}
\usage{
cache_get_partitioning(x)
}
\arguments{
\item{x}{Arrow Dataset}
}
\value{
a named vector, the key is the primary_key, the value is an un-evaluated R expression to be used in creating the partitioning by
}
\description{
Internal function to read/generate partition information from an Arrow Dataset.
}
\examples{
\dontrun{
cache_get_partitioning(cache_read("test","deep","stream"))
}
}
