% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_make_partitioning.R
\name{cache_make_partitioning}
\alias{cache_make_partitioning}
\title{cache_make_partitioning}
\usage{
cache_make_partitioning(x, primary_keys = attr(x, "primary_keys"))
}
\arguments{
\item{x}{data.frame}

\item{primary_keys}{primary keys to set partitioning with}
}
\value{
an unevaluated expression determining the partitioning
}
\description{
Builds suggested partitioning from the primary keys of the table.
}
\examples{
x <- data.frame(a = c(1, 2, 3), b = c("a", "b", "c"))

cache_make_partitioning(x, primary_keys = "a")
# expression a/10000
}
