% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_files.R
\name{cache_path}
\alias{cache_path}
\title{cache_path}
\usage{
cache_path(
  table_name,
  depth = c("deep", "shallow"),
  type = c("tessi", "stream")
)
}
\arguments{
\item{table_name}{string}

\item{depth}{string, either "deep" or "shallow"}

\item{type}{string, either "tessi" or "stream"}
}
\value{
string for the configured cache path
}
\description{
Internal function to build cache directory/path and check that we have read/write access
\verb{tessilake.\{depth\}/\{type\}/\{table_name\}}
}
\examples{
\dontrun{
cache_path("test", "deep", "stream")
}
}
