% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_read}
\alias{cache_read}
\title{cache_read}
\usage{
cache_read(
  table_name,
  depth = c("deep", "shallow"),
  type = c("tessi", "stream"),
  include_partition = FALSE,
  ...
)
}
\arguments{
\item{table_name}{string}

\item{depth}{string, either "deep" or "shallow"}

\item{type}{string, either "tessi" or "stream"}

\item{include_partition}{boolean, whether or not to return the partition information as a column}

\item{...}{extra arguments to pass on to arrow::open_dataset}
}
\value{
\code{\link[arrow:Dataset]{arrow::Dataset}}
}
\description{
Internal function to read cached arrow files. Simple wrapper on open_dataset.
Optionally returns the partition information as a dataset column.
}
\examples{
\dontrun{
cache_read("test","deep","stream")
}
}
