% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_attributes.R
\name{cache_set_attributes}
\alias{cache_set_attributes}
\title{cache_set_attributes}
\usage{
cache_set_attributes(x, attributes)
}
\arguments{
\item{x}{Arrow Object}

\item{attributes}{list of attributes to set}
}
\description{
Internal function to set R attributes from an Arrow Dataset/arrow_dplyr_query
}
\examples{
\dontrun{
cache_set_attributes(cache_read("test", "deep", "stream"), list(primary_keys = "a"))
}
}
