% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_update.R
\name{cache_update}
\alias{cache_update}
\title{cache_update}
\usage{
cache_update(
  x,
  table_name,
  depth = c("deep", "shallow"),
  type = c("tessi", "stream"),
  primary_keys = attr(x, "primary_keys"),
  date_column = NULL,
  delete = FALSE,
  ...
)
}
\arguments{
\item{x}{data.frame or part of a data.frame to be cached}

\item{table_name}{string}

\item{depth}{string, either "deep" or "shallow"}

\item{type}{string, either "tessi" or "stream"}

\item{primary_keys}{character vector of columns to be used for partitioning, only the first one is currently used}

\item{date_column}{character name of the column to be used for determining the date of last row update}

\item{delete}{whether to delete rows in cache missing from \code{x}, default is not to delete the rows}

\item{...}{extra arguments passed on to \code{\link[arrow:open_dataset]{arrow::open_dataset}} and \code{\link[arrow:write_dataset]{arrow::write_dataset}}}
}
\value{
invisible
}
\description{
Internal function to update portions of cached arrow files.
}
\examples{
\dontrun{
x <- data.table(a = 1:1000, b = runif(1000))
y <- data.table(b = 100:199, b = runif(100))
cache_write(x, "test", "deep", "stream")
cache_update(y, "test", "deep", "stream")
}
}
