% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_write}
\alias{cache_write}
\title{cache_write}
\usage{
cache_write(
  x,
  table_name,
  depth = c("deep", "shallow"),
  type = c("tessi", "stream"),
  primary_keys = attr(x, "primary_keys"),
  partition = !is.null(primary_keys),
  ...
)
}
\arguments{
\item{x}{data.frame to be written}

\item{table_name}{string}

\item{depth}{string, either "deep" or "shallow"}

\item{type}{string, either "tessi" or "stream"}

\item{primary_keys}{character vector of columns to be used for partitioning, only the first one is currently used}

\item{partition}{boolean, whether or not to partition the dataset using primary_keys information}

\item{...}{extra arguments passed on to \code{\link[arrow:write_dataset]{arrow::write_dataset}}}
}
\value{
invisible
}
\description{
Internal function to write cached arrow files. Simple wrapper on write_dataset that points to a directory defined by
\verb{tessilake.\{depth\}/\{type\}} and uses partitioning specified by primary_keys attribute/argument.
}
\examples{
\dontrun{
x = data.table(a=c(1,2,3))
cache_write(x,"test","deep","stream",primary_keys=c("a"))
}
}
