% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-error.R
\name{is.error}
\alias{is.error}
\title{is.error}
\usage{
is.error(expr)
}
\arguments{
\item{expr}{expression to run (in the parent frame)}
}
\value{
boolean, TRUE if the expression causes an error, FALSE if not
}
\description{
Helper function to use error-throwing expressions in stopifnot
}
\examples{
is.error(stop("Error"))

}
