% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_readwrite.R
\name{read_cache}
\alias{read_cache}
\alias{cache_read}
\title{read_cache}
\usage{
read_cache(table_name, type, depth = deprecated(), ...)

cache_read(
  table_name,
  depth,
  type,
  include_partition = FALSE,
  select = NULL,
  num_tries = 60,
  ...
)
}
\arguments{
\item{table_name}{string}

\item{type}{string, e.g. "tessi" or "stream"}

\item{depth}{string, e.g. "deep" or "shallow", deprecated in \link{read_cache}}

\item{...}{extra arguments to pass on to \link{cache_read} and then \link[arrow:open_dataset]{arrow::open_dataset}, \link[arrow:read_parquet]{arrow::read_parquet}, or \link[arrow:read_feather]{arrow::read_feather}}

\item{include_partition}{boolean, whether or not to return the partition information as a column}

\item{select}{vector of strings indicating columns to select from database}

\item{num_tries}{integer number of times to try reading before failing}
}
\value{
\link[arrow:Dataset]{arrow::Dataset}
}
\description{
Function to read cached \link[arrow:arrow-package]{arrow} files. Reads from the most recently updated cache across all storage depths.
Optionally returns the partition information as a dataset column.
}
\section{Functions}{
\itemize{
\item \code{cache_read()}: Underlying cache reader that invokes \link[arrow:open_dataset]{arrow::open_dataset}, \link[arrow:read_parquet]{arrow::read_parquet}, or \link[arrow:read_feather]{arrow::read_feather}
and handles partitioning information

}}
\examples{
\dontrun{
write_cache(data.table(a=letters), "test", "stream")
read_cache("test", "stream")
}
}
