% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_readwrite.R
\name{read_cache}
\alias{read_cache}
\title{read_cache}
\usage{
read_cache(
  table_name,
  depth,
  type,
  include_partition = FALSE,
  select = NULL,
  num_tries = 60,
  ...
)
}
\arguments{
\item{table_name}{string}

\item{depth}{string, e.g. "deep" or "shallow"}

\item{type}{string, e.g. "tessi" or "stream"}

\item{include_partition}{boolean, whether or not to return the partition information as a column}

\item{select}{vector of strings indicating columns to select from database}

\item{num_tries}{integer number of times to try reading before failing}

\item{...}{extra arguments to pass on to arrow::open_dataset}
}
\value{
\code{\link[arrow:Dataset]{arrow::Dataset}}
}
\description{
Function to read cached arrow files. Simple wrapper on open_dataset.
Optionally returns the partition information as a dataset column.
}
\examples{
\dontrun{
read_cache("test", "deep", "stream")
}
}
