% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{read_sql}
\alias{read_sql}
\alias{read_sql_table}
\title{read_sql}
\usage{
read_sql(
  query,
  name = digest::sha1(query),
  select = NULL,
  primary_keys = NULL,
  date_column = NULL,
  freshness = as.difftime(7, units = "days"),
  incremental = TRUE
)

read_sql_table(
  table_name,
  schema = "dbo",
  select = NULL,
  primary_keys = NULL,
  date_column = NULL,
  freshness = as.difftime(7, units = "days"),
  incremental = TRUE
)
}
\arguments{
\item{query}{character query to run on the database.}

\item{name}{name of the query, defaults to the SHA1 hash of the query.}

\item{select}{vector of strings indicating columns to select from database}

\item{primary_keys}{character vector, primary keys of the table. \code{read_sql_table} will attempt to identify the primary keys using
SQL metadata.}

\item{date_column}{character, date column of the table showing the last date the row was updated.
Defaults to "last_update_dt" if it exists in the table.}

\item{freshness}{the returned data will be at least this fresh}

\item{incremental}{whether or not to load data incrementally, default is \code{TRUE}}

\item{table_name}{character, table name without schema.}

\item{schema}{character, database schema. Default is \code{dbo}}
}
\value{
an Apache Arrow Table, see the \link[arrow:arrow-package]{arrow::arrow-package} package for more information.

an Apache Arrow Table, see the \link[arrow:arrow-package]{arrow::arrow-package} package for more information.
}
\description{
Execute a database query and cache it locally as a Feather file and remotely as a Parquet file.
Cache storage locations are managed by \code{tessilake.depths} options.
Database connection defined by an ODBC profile with the name set by the \code{tessilake.tessitura} option.
}
\section{Functions}{
\itemize{
\item \code{read_sql_table()}: Reads a table or view from a SQL database and caches it locally using read_sql.

}}
\examples{
\dontrun{
read_sql("select * from T_CUSTOMER", "t_customer",
  primary_keys = "customer_no",
  date_column = "last_update_dt"
)
}
\dontrun{
read_sql_table("T_CUSTOMER")
}
}
