% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessi.R
\name{read_tessi}
\alias{read_tessi}
\title{read_tessi}
\usage{
read_tessi(
  table_name,
  select = NULL,
  freshness = as.difftime(7, units = "days"),
  incremental = TRUE,
  ...
)
}
\arguments{
\item{table_name}{character name of the table to read from Tessitura, either one of the available tables (see \code{\link[=tessi_list_tables]{tessi_list_tables()}}) or the
name of a SQL table that exists in Tessitura. The default SQL table schema is \code{dbo}.}

\item{select}{vector of strings indicating columns to select from database}

\item{freshness}{the returned data will be at least this fresh}

\item{incremental}{whether or not to load data incrementally, default is \code{TRUE}}

\item{...}{further arguments to be passed to read_sql_table}
}
\value{
an Apache \link[arrow:Table]{arrow::Table}.
}
\description{
Thin wrapper on read_sql_table using the tables configured in \code{list_tessi_tables} that also :
\itemize{
\item updates \code{customer_no} based on merges and adds \code{group_customer_no} based on \code{\link[=tessi_customer_no_map]{tessi_customer_no_map()}}
}
}
\examples{
\dontrun{
read_tessi("memberships",
  select = c("memb_level", "customer_no")
)
}

}
