% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-table.R
\name{setleftjoin}
\alias{setleftjoin}
\title{setleftjoin
Fast left join for data.table that simply adds r's columns to l by reference.}
\usage{
setleftjoin(l, r, by = NULL)
}
\arguments{
\item{l}{left side of the join (data.table)}

\item{r}{right side of the join (data.table)}

\item{by}{character vector for matching columns. Names, if they exist, refer to columns in the
left table, values to columns in the right table}
}
\value{
a joined data.table, invisibly
}
\description{
setleftjoin
Fast left join for data.table that simply adds r's columns to l by reference.
}
\examples{

library(data.table)

l = data.table(a=c(1,2,3),b=c(3,4,NA))
r = data.table(b=c(3,7,NA),c=c(1,2,4))

setleftjoin(l,r,by="b")
setleftjoin(l,r,by=c("b"="c"))

}
