% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessi.R
\docType{data}
\name{tessi_list_tables}
\alias{tessi_list_tables}
\alias{tessi_tables}
\title{tessi_list_tables}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 52 rows and 4 columns.
}
\usage{
tessi_list_tables()

tessi_tables
}
\value{
data.table of configured tessitura tables with columns short_name, long_name, baseTable and primary_keys
}
\description{
The list of tessitura tables is configured in the extdata/tessi_tables.yml file in the package directory
}
\details{
\subsection{yml format}{\preformatted{\{short_name\}:
   long_name: \{name of table/view to be loaded\}
   baseTable: \{the underlying table being queried that has primary_keys\}
   primary_keys: \{the primary key(s) as a value or a list of values\}
}
}
}
\examples{
# customers:
#   long_name: BI.VT_CUSTOMER
#   baseTable: T_CUSTOMER
#   primary_keys: customer_no

list_tessi_tables()[short_name == "customers"]

}
\keyword{datasets}
