% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-table.R
\name{update_table}
\alias{update_table}
\title{update_table}
\usage{
update_table(from, to, dateColumn = NULL, primaryKeys = NULL)
}
\arguments{
\item{from}{the source object}

\item{to}{the object to be updated}

\item{dateColumn}{string or tidyselected column identifying the date column to use to determine which rows to update}

\item{primaryKeys}{vector of strings or tidyselected columns identifying the primary keys to determine which rows to update}
}
\value{
an updated data.table updated in-place
}
\description{
updates the data.table-like object \code{to} based on the data.frame-like \code{from} object,
using \code{dateColumn} to determine which rows have been updated and updating only those with matching \code{primaryKeys}.
}
\examples{
from = data.frame()


}
