% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_table.R
\name{update_table}
\alias{update_table}
\alias{update_table.default}
\alias{update_table.data.table}
\title{update_table}
\usage{
update_table(
  from,
  to,
  date_column = NULL,
  primary_keys = NULL,
  delete = FALSE,
  incremental = TRUE
)

\method{update_table}{default}(
  from,
  to,
  date_column = NULL,
  primary_keys = NULL,
  delete = FALSE,
  incremental = TRUE
)

\method{update_table}{data.table}(
  from,
  to,
  date_column = NULL,
  primary_keys = NULL,
  delete = FALSE,
  incremental = TRUE
)
}
\arguments{
\item{from}{the source object}

\item{to}{the object to be updated}

\item{date_column}{string or tidyselected column identifying the date column to use to determine which rows to update}

\item{primary_keys}{vector of strings or tidyselected columns identifying the primary keys to determine which rows to update}

\item{delete}{whether to delete rows in \code{to} missing from \code{from}, default is not to delete the rows}

\item{incremental}{whether or not to update the table incrementally or to simply overwrite the existing table with a new one,
the default is \code{TRUE}}
}
\value{
a data.table updated in-place
}
\description{
updates the data.table-like object \code{to} based on the data.frame-like \code{from} object,
using \code{date_column} to determine which rows have been updated and updating only those with matching \code{primary_keys}.
}
\examples{
library(data.table)
expect <- data.table(expand.grid(x = 1:100, y = 1:100))[, data := runif(.N)]
# divide the data
from <- copy(expect)[1:9000]
to <- copy(expect)[1000:10000]
# and mung it up
to[1:5000, data := runif(.N)]

update_table(from, to, primary_keys = c(x, y)) == expect
# TRUE

}
