% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_readwrite.R
\name{write_cache}
\alias{write_cache}
\alias{cache_write}
\title{write_cache}
\usage{
write_cache(
  x,
  table_name,
  type,
  depth = deprecated(),
  incremental = FALSE,
  ...
)

cache_write(
  x,
  table_name,
  depth,
  type,
  primary_keys = cache_get_attributes(x)$primary_keys,
  partition = !is.null(primary_keys),
  overwrite = FALSE,
  num_tries = 60,
  ...
)
}
\arguments{
\item{x}{data.frame to be written}

\item{table_name}{string}

\item{type}{string, e.g. "tessi" or "stream"}

\item{depth}{string, e.g. "deep" or "shallow", deprecated in \link{write_cache}}

\item{incremental}{boolean, whether to call \link{cache_update} or \link{cache_write} to update the cached dataset.}

\item{...}{extra arguments passed on to \link[arrow:write_feather]{arrow::write_feather}, \link[arrow:write_parquet]{arrow::write_parquet} or \link[arrow:write_dataset]{arrow::write_dataset}}

\item{primary_keys}{character vector of columns to be used for partitioning, only the first one is currently used}

\item{partition}{boolean, whether or not to partition the dataset using primary_keys information}

\item{overwrite}{boolean, whether or not to overwrite an existing cache}

\item{num_tries}{integer number of times to try reading before failing}
}
\value{
invisible
}
\description{
Function to write cached \link[arrow:arrow-package]{arrow} files. Writes to the first defined storage in \code{config::get("tessilake")},
and then syncs to the other locations by calling \link{sync_cache}.
}
\section{Functions}{
\itemize{
\item \code{cache_write()}: Underlying cache writer that invokes \link[arrow:write_feather]{arrow::write_feather}, \link[arrow:write_parquet]{arrow::write_parquet} or \link[arrow:write_dataset]{arrow::write_dataset} and handles partitioning
specified by \code{primary_keys} attribute/argument.

}}
\examples{
\dontrun{
x <- data.table(a = c(1, 2, 3))
write_cache(x, "test", "stream", primary_keys = c("a"))
}
}
