% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ff.R
\name{write_ffdf}
\alias{write_ffdf}
\title{write_ffdf}
\usage{
write_ffdf(x, table_name, out_dir)
}
\arguments{
\item{x}{data.frame to be written}

\item{table_name}{string name of the file to be written, will be converted to camelcase}

\item{out_dir}{string directory where the gzipped ffdf should be written}
}
\value{
invisibly
}
\description{
Write to gzipped ffdf. \code{table_name} is converted to camelcase for consistency with
legacy use
}
